test deploy 5

+++ Jenkinsfile
@@ -12,13 +12,15 @@ node {
             sh 'py.test --junit-xml test-reports/results.xml sources/test_calc.py'
             junit 'test-reports/results.xml'
         }
-        withEnv(['VOLUME = \'$(pwd)/sources:/src\'', 'IMAGE = \'cdrx/pyinstaller-linux:python2\'']) {
+        withEnv(['VOLUME=\'$(pwd)/sources:/src\'', 'IMAGE=\'cdrx/pyinstaller-linux:python2\'']) {
             stage('Deploy') {
                 dir('env.BUILD_ID') {
                     unstash name: 'compiled-results'
                     sh "docker run --rm -v ${VOLUME} ${IMAGE} 'pyinstaller -F add2vals.py'"
-                    archiveArtifacts artifacts: '${env.BUILD_ID}/sources/dist/add2vals', followSymlinks: false
-                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
+                    if (currentBuild.result == null || currentBuild.result == 'SUCCESS'){
+                        archiveArtifacts artifacts: '${env.BUILD_ID}/sources/dist/add2vals', followSymlinks: false
+                        sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
+                    }
                 }
             }
         }
